Support reading multiple files from URL instead of a single file (#387)
**User-Facing Changes**
Lichtblick users will now be able to open multiple files on the web
version just like in the desktop version. For now this will be only
doable via URL, for instance creating a script that joins multiple
sources into a single URL and then pushing it to a browser.
A link to accessing multiple files would look something like this: 


`http://example.com/?ds=remote-file&ds.url=http%3A%2F%2Flichtblick.dev%2Ftest-file1.mcap&ds.url=http%3A%2F%2Flichtblick.dev%2Ftest-file2.mcap`

To note that there are some restrictions for this to be able to be used,
the sources must point to a .mcap file and all the files need to be
.mcap

**Description**

Changed the way the application handles MCAP data sources from URL to
now accept an array with all the sources passed. This makes it simpler
to deal one or multiple sources.

Note: This PR is pointing to be merged to
[story/load-multiple-mcaps-r1](https://github.com/Lichtblick-Suite/lichtblick/tree/story/load-multiple-mcaps-r1)
branch just for facilitating the review (it was created from this branch
to be able to access features that aren't implemented on the main branch
yet), this will be merged directly to main but only after the branch
story is merged.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [X] The web version was tested and it is running ok
- [X] The desktop version was tested and it is running ok
- [X] This change is covered by unit tests
- [X] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated

---------

Signed-off-by: ctw-joao-luis <Joao.Luis@ctw.bmwgroup.com>
Co-authored-by: Luiz Bezerra <luiz.bezerra@ctw.bmwgroup.com>
Co-authored-by: Alexandre Neuwald CTW <Alexandre.Neuwald@ctw.bmwgroup.com>
Co-authored-by: Alexandre Neuwald <alexandre.neuwald@criticaltechworks.com>