fix: call message converters when the settings change
Adds a new input to the buildRenderState() function called `forceConversion`. As
the name implies, any topic in this set will be forcefully reconverted on the
current frame.

Right now, its only being used after the callback for a given message converter,
but it could be a useful facility to expose to the user in the render state at
some point.

Should fix #503.